backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
openai
python-dotenv
pydantic

